+++ iGCS/Comm/Interfaces/Bluetooth/BluetoothStream.m
@@ -48,7 +48,7 @@
 
 
 // MavLinkInterface destination override (used for Tx)
--(void)consumeData:(const uint8_t*)bytes length:(int)length {
+-(void)consumeData:(const uint8_t*)bytes length:(NSInteger)length {
     if (self.gkLocalController) {
         [self.gkLocalController sendMavlinkData:bytes length:length];
     }
