+++ iGCS/UserInterface/Annotations/WaypointAnnotation.m
@@ -12,7 +12,7 @@
 
 @implementation WaypointAnnotation
 
-- (id)initWithCoordinate:(CLLocationCoordinate2D)coordinate andWayPoint:(mavlink_mission_item_t)waypoint atIndex:(int)index {
+- (id)initWithCoordinate:(CLLocationCoordinate2D)coordinate andWayPoint:(mavlink_mission_item_t)waypoint atIndex:(NSInteger)index {
     if ((self = [super init])) {
         _coordinate = coordinate;
         _waypoint   = waypoint;
@@ -52,7 +52,7 @@
     return [theme waypointOtherColor];
 }
 
-- (BOOL) isCurrentWaypointP:(int)currentWaypointSeq {
+- (BOOL) isCurrentWaypointP:(NSInteger)currentWaypointSeq {
     return (currentWaypointSeq == _waypoint.seq);
 }
